[PSEC-1488] Audit EC2 instances
Co-authored-by: Freshrojek <jnboyer1@gmail.com>

A "quick and dirty" implementation for this ticket was requested
urgently before a Risk Review meeting as the outcome of the EC2
instances audit was needed for a Platform Team to progress on their risk
ticket (RISK-46).

The logic added in this commit needs to be covered by tests and could probably
do with a little bit of clean-up / refactor.